requirements.txt
# Core dependencies for Flat Syntax annotation pipeline
requests>=2.31.0
click>=8.1.0
pyyaml>=6.0

# Optional: Cloud LLM providers (for comparison/benchmarking only)
# Uncomment if needed for testing against cloud models
# openai>=1.0.0
# anthropic>=0.7.0

# Development dependencies (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
